+++ src/components/LoginForm.vue
@@ -1,0 +1,90 @@
+<template>
+  <v-card
+    class="elevation-12">
+    <v-toolbar dark color="primary">
+      <v-toolbar-title>
+        Sign in
+      </v-toolbar-title>
+    </v-toolbar>
+    <v-card-text>
+      <loader v-if="isLoading" />
+      <v-form v-else>
+        <v-text-field
+          v-model="email"
+          id="email"
+          name="email"
+          label="Email"
+          type="email"
+        ></v-text-field>
+        <v-text-field
+          v-model="password"
+          id="password"
+          name="password"
+          label="Password"
+          type="password"
+        ></v-text-field>
+      </v-form>
+      <v-alert
+        v-if="getErrorMessage"
+        border="top"
+        color="red lighten-2"
+        dark
+      >
+        {{ getErrorMessage }}
+      </v-alert>
+    </v-card-text>
+    <v-card-actions>
+      <v-spacer></v-spacer>
+      <v-btn
+        @click="onSubmit"
+        :disabled="isLoading"
+        color="primary">
+        Login
+      </v-btn>
+    </v-card-actions>
+  </v-card>
+</template>
+
+<script>
+import { mapActions, mapGetters, mapState } from 'vuex';
+import { PROFILE_PAGE } from '@/config/constants';
+import Loader from '@/components/Loader';
+
+export default {
+  name: 'LoginForm',
+  components: {
+    Loader,
+  },
+  data() {
+    return {
+      email: '',
+      password: '',
+    };
+  },
+  computed: {
+    ...mapState('loginPage', [
+      'error',
+      'isLoading',
+    ]),
+    ...mapGetters('loginPage', [
+      'getErrorMessage',
+    ]),
+  },
+  methods: {
+    async onSubmit() {
+      const result = await this.checkUser({ email: this.email, password: this.password });
+
+      if (result) {
+        await this.$router.push({ path: PROFILE_PAGE });
+      }
+    },
+    ...mapActions('loginPage', [
+      'checkUser',
+    ]),
+  },
+};
+</script>
+
+<style scoped>
+
+</style>
