+++ src/views/LoginView.vue
@@ -5,91 +5,19 @@
         xs12
         sm8
         md4 >
-        <v-card
-          class="elevation-12">
-          <v-toolbar dark color="primary">
-            <v-toolbar-title>
-              Sign in
-            </v-toolbar-title>
-          </v-toolbar>
-          <v-card-text>
-            <loader v-if="isLoading" />
-            <v-form v-else>
-              <v-text-field
-                v-model="email"
-                id="email"
-                name="email"
-                label="Email"
-                type="email"
-              ></v-text-field>
-              <v-text-field
-                v-model="password"
-                id="password"
-                name="password"
-                label="Password"
-                type="password"
-              ></v-text-field>
-            </v-form>
-            <v-alert
-              v-if="getErrorMessage"
-              border="top"
-              color="red lighten-2"
-              dark
-            >
-              {{ getErrorMessage }}
-            </v-alert>
-          </v-card-text>
-          <v-card-actions>
-            <v-spacer></v-spacer>
-            <v-btn
-              @click="onSubmit"
-              :disabled="isLoading"
-              color="primary">
-              Login
-            </v-btn>
-          </v-card-actions>
-        </v-card>
+        <login-form />
       </v-flex>
     </v-layout>
   </v-container>
 </template>
 
 <script>
-import { mapActions, mapGetters, mapState } from 'vuex';
-import Loader from '@/components/Loader';
-import { PROFILE_PAGE } from '@/config/constants';
+import LoginForm from '@/components/LoginForm';
 
 export default {
   name: 'LoginView',
   components: {
-    Loader,
-  },
-  data() {
-    return {
-      email: '',
-      password: '',
-    };
-  },
-  computed: {
-    ...mapState('loginPage', [
-      'error',
-      'isLoading',
-    ]),
-    ...mapGetters('loginPage', [
-      'getErrorMessage',
-    ]),
-  },
-  methods: {
-    async onSubmit() {
-      const result = await this.checkUser({ email: this.email, password: this.password });
-
-      if (result) {
-        await this.$router.push({ path: PROFILE_PAGE });
-      }
-    },
-    ...mapActions('loginPage', [
-      'checkUser',
-    ]),
+    LoginForm,
   },
 };
 </script>
